Make `:prefix` description a bullet item
The README contains the odd-looking heading "tem", which looks like it
was meant to be "item" and hence actually a bullet item like the rest of
the items explaining the options to `wrap-text`.  This change turns the
description of `:prefix` into a bullet item to be consistent with the
rest of the text.